# customers/src/main/resources/static/openapi/customers-api.yml
openapi: 3.0.1
info:
  title: Customers API
  description: API for managing customers
  version: 1.0.0
  contact:
    name: Reymon Hidalgo
paths:
  /api/v1/customers:
    post:
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerDTO'
      responses:
        '201':
          description: Customer created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponseGetCustomerDTO'
    get:
      summary: Get all customers
      parameters:
        - name: state
          in: query
          required: true
          schema:
            type: boolean
        - name: page
          in: query
          required: true
          schema:
            type: integer
        - name: size
          in: query
          required: true
          schema:
            type: integer
        - name: sorting
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Customers retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponsePageGetCustomerDTO'
  /api/v1/customers/{identificationNumber}:
    get:
      summary: Get a customer by identification number
      parameters:
        - name: identificationNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Customer retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponseGetCustomerDTO'
    put:
      summary: Update a customer
      parameters:
        - name: identificationNumber
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerDTO'
      responses:
        '200':
          description: Customer updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponseGetCustomerDTO'
    delete:
      summary: Delete a customer
      parameters:
        - name: identificationNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Customer deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenericResponseVoid'
components:
  schemas:
    CreateCustomerDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        password:
          type: string
        state:
          type: boolean
        person:
          $ref: '#/components/schemas/CreatePersonDTO'
    CreatePersonDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        identificationNumber:
          type: string
        gender:
          type: string
        address:
          type: string
        phone:
          type: string
    GetCustomerDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        state:
          type: boolean
        customerName:
          type: string
        phone:
          type: string
        address:
          type: string
        identificationNumber:
          type: string
    GenericResponseGetCustomerDTO:
      type: object
      properties:
        success:
          type: boolean
        status:
          type: integer
        message:
          type: string
        path:
          type: string
        data:
          $ref: '#/components/schemas/GetCustomerDTO'
    GenericResponsePageGetCustomerDTO:
      type: object
      properties:
        success:
          type: boolean
        status:
          type: integer
        message:
          type: string
        path:
          type: string
        data:
          type: object
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/GetCustomerDTO'
    GenericResponseVoid:
      type: object
      properties:
        success:
          type: boolean
        status:
          type: integer
        message:
          type: string
        path:
          type: string
        data:
          type: object
          nullable: true